Rename dev-server.sj to main.js
Refs #109.
